Add files via upload
This PDF contains the final manuscript of my undergraduate thesis, “Does Money Matter? Examining the Relationships Between Federal Funding & Population Trends for ESA-Listed Marine Species.” The thesis presents an analysis of the impact of federal funding on the population trends of marine species listed under the Endangered Species Act (ESA). It includes detailed sections on data collection, statistical methodology, results, and a discussion of the implications for conservation policy.